re-arrange settings files
