+++ server_addon/houdini/server/settings/create.py
@@ -1,5 +1,4 @@
 from pydantic import Field
-
 from ayon_server.settings import BaseSettingsModel
 
 
@@ -37,7 +36,7 @@ class CreateStaticMeshModel(BaseSettingsModel):
 class CreatePluginsModel(BaseSettingsModel):
     CreateArnoldAss: CreateArnoldAssModel = Field(
         default_factory=CreateArnoldAssModel,
-        title="Create Alembic Camera")
+        title="Create Arnold Ass")
     # "-" is not compatible in the new model
     CreateStaticMesh: CreateStaticMeshModel = Field(
         default_factory=CreateStaticMeshModel,
@@ -134,86 +133,4 @@ DEFAULT_HOUDINI_CREATE_SETTINGS = {
         "enabled": False,
         "default_variants": ["Main"]
     },
-}
-
-
-# Publish Plugins
-class ValidateWorkfilePathsModel(BaseSettingsModel):
-    enabled: bool = Field(title="Enabled")
-    optional: bool = Field(title="Optional")
-    node_types: list[str] = Field(
-        default_factory=list,
-        title="Node Types"
-    )
-    prohibited_vars: list[str] = Field(
-        default_factory=list,
-        title="Prohibited Variables"
-    )
-
-
-class BasicValidateModel(BaseSettingsModel):
-    enabled: bool = Field(title="Enabled")
-    optional: bool = Field(title="Optional")
-    active: bool = Field(title="Active")
-
-
-class PublishPluginsModel(BaseSettingsModel):
-    ValidateWorkfilePaths: ValidateWorkfilePathsModel = Field(
-        default_factory=ValidateWorkfilePathsModel,
-        title="Validate workfile paths settings.")
-    ValidateReviewColorspace: BasicValidateModel = Field(
-        default_factory=BasicValidateModel,
-        title="Validate Review Colorspace.")
-    ValidateContainers: BasicValidateModel = Field(
-        default_factory=BasicValidateModel,
-        title="Validate Latest Containers.")
-    ValidateSubsetName: BasicValidateModel = Field(
-        default_factory=BasicValidateModel,
-        title="Validate Subset Name.")
-    ValidateMeshIsStatic: BasicValidateModel = Field(
-        default_factory=BasicValidateModel,
-        title="Validate Mesh is Static.")
-    ValidateUnrealStaticMeshName: BasicValidateModel = Field(
-        default_factory=BasicValidateModel,
-        title="Validate Unreal Static Mesh Name.")
-
-
-DEFAULT_HOUDINI_PUBLISH_SETTINGS = {
-    "ValidateWorkfilePaths": {
-        "enabled": True,
-        "optional": True,
-        "node_types": [
-            "file",
-            "alembic"
-        ],
-        "prohibited_vars": [
-            "$HIP",
-            "$JOB"
-        ]
-    },
-    "ValidateReviewColorspace": {
-        "enabled": True,
-        "optional": True,
-        "active": True
-    },
-    "ValidateContainers": {
-        "enabled": True,
-        "optional": True,
-        "active": True
-    },
-    "ValidateSubsetName": {
-        "enabled": True,
-        "optional": True,
-        "active": True
-    },
-    "ValidateMeshIsStatic": {
-        "enabled": True,
-        "optional": True,
-        "active": True
-    },
-    "ValidateUnrealStaticMeshName": {
-        "enabled": False,
-        "optional": True,
-        "active": True
-    }
 }
