+++ server_addon/houdini/server/settings/main.py
@@ -1,57 +1,19 @@
 from pydantic import Field
-from ayon_server.settings import (
-    BaseSettingsModel,
-    MultiplatformPathModel,
-    MultiplatformPathListModel,
-)
+from ayon_server.settings import BaseSettingsModel
 from .general import (
     GeneralSettingsModel,
     DEFAULT_GENERAL_SETTINGS
 )
 from .imageio import HoudiniImageIOModel
-from .publish_plugins import (
-    PublishPluginsModel,
+from .shelves import ShelvesModel
+from .create import (
     CreatePluginsModel,
-    DEFAULT_HOUDINI_PUBLISH_SETTINGS,
     DEFAULT_HOUDINI_CREATE_SETTINGS
 )
-
-
-class ShelfToolsModel(BaseSettingsModel):
-    name: str = Field(title="Name")
-    help: str = Field(title="Help text")
-    script: MultiplatformPathModel = Field(
-        default_factory=MultiplatformPathModel,
-        title="Script Path "
-    )
-    icon: MultiplatformPathModel = Field(
-        default_factory=MultiplatformPathModel,
-        title="Icon Path "
-    )
-
-
-class ShelfDefinitionModel(BaseSettingsModel):
-    _layout = "expanded"
-    shelf_name: str = Field(title="Shelf name")
-    tools_list: list[ShelfToolsModel] = Field(
-        default_factory=list,
-        title="Shelf Tools"
-    )
-
-
-class ShelvesModel(BaseSettingsModel):
-    _layout = "expanded"
-    shelf_set_name: str = Field(title="Shelfs set name")
-
-    shelf_set_source_path: MultiplatformPathListModel = Field(
-        default_factory=MultiplatformPathListModel,
-        title="Shelf Set Path (optional)"
-    )
-
-    shelf_definition: list[ShelfDefinitionModel] = Field(
-        default_factory=list,
-        title="Shelf Definitions"
-    )
+from .publish import (
+    PublishPluginsModel,
+    DEFAULT_HOUDINI_PUBLISH_SETTINGS,
+)
 
 
 class HoudiniSettings(BaseSettingsModel):
@@ -65,17 +27,15 @@ class HoudiniSettings(BaseSettingsModel):
     )
     shelves: list[ShelvesModel] = Field(
         default_factory=list,
-        title="Houdini Scripts Shelves",
+        title="Shelves Manager",
     )
-
-    publish: PublishPluginsModel = Field(
-        default_factory=PublishPluginsModel,
-        title="Publish Plugins",
-    )
-
     create: CreatePluginsModel = Field(
         default_factory=CreatePluginsModel,
         title="Creator Plugins",
+    )
+    publish: PublishPluginsModel = Field(
+        default_factory=PublishPluginsModel,
+        title="Publish Plugins",
     )
