+++ server_addon/houdini/server/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.4"
+__version__ = "0.1.5"
